Add files via upload
Cargo is a file that is a configuration file used by Cargo, the package manager and build system for Rust. It plays a central role in managing and defining the properties of a Rust project and in this rust project, the Cargo.toml file creates the metadata, adds dependencies, imports binary executables, and provides workspace configuration. It also provides the randomness of the users and the emails that fill the array in the "mainArray.rs" Rust file.